feat(lab2): created classes for tables and a database handler, added database loading

+++ KBD_2/src/sample/database/equipment.sql
@@ -3,5 +3,5 @@ CREATE TABLE equipment
     inventory_number INT PRIMARY KEY,
     name             VARCHAR(30),
     date_of_purchase DATE,
-    price            INT
+    price            DECIMAL
 );
